Less explicit, but more correct?
