+++ src/server/frontend/index.js
@@ -5,6 +5,7 @@ import express from 'express';
 // import favicon from 'serve-favicon';
 import render from './render';
 import userState from './userstate';
+import i18nLoader from '../lib/i18nmiddleware';
 
 const app = express();
 
@@ -19,7 +20,8 @@ app.use(compression());
 app.use('/build', express.static('build'));
 app.use('/assets', express.static('assets'));
 
-// Load state extras for current user.
+// Load state extras for current user
+app.use(i18nLoader(config.defaultLocale));
 app.use(userState());
 
 app.get('*', (req, res, next) => {
