+++ src/server/frontend/userstate.js
@@ -1,7 +1,5 @@
-import config from '../config';
 import Promise from 'bluebird';
 import Immutable from 'immutable';
-import i18n from '../lib/loadtranslations';
 
 // Example how initialState, which is the same for all users, is enriched with
 // user state. With state-less Flux, we don't need instances.
@@ -9,8 +7,7 @@ export default function userState() {
 
   return (req, res, next) => {
     loadUserData(req).then(loadedData => {
-      console.log(loadedData);
-      req.userState = Immutable.Map().merge(...loadedData);
+      req.userState = Immutable.Map().merge(...loadedData, req.i18n);
       next();
     });
   };
@@ -20,7 +17,6 @@ export default function userState() {
 // Gracefully settle all promises, ignore failed.
 function loadUserData(req) {
   const dataSources = [
-    loadLanguage(req),
     loadTodos()
   ];
 
@@ -29,17 +25,6 @@ function loadUserData(req) {
       .filter(promise => promise.isFulfilled())
       .map(promise => promise.value())
   );
-}
-
-function loadLanguage(req) {
-  const acceptedLanguage = req.acceptsLanguages(i18n.locales);
-  const locale = acceptedLanguage || config.defaultLocale;
-  return {
-    i18n: {
-      locale,
-      messages: i18n.messages[locale]
-    }
-  };
 }
 
 // Simulate async action.
